+++ avian3d/src/lib.rs
@@ -84,9 +84,28 @@ impl Plugin for TnuaAvian3dPlugin {
             self.schedule,
             apply_motors_system.in_set(TnuaPipelineStages::Motors),
         );
+        app.add_systems(
+            Update,
+            ensure_subservient_sensors_are_colliders_of_their_bevy_parents,
+        );
         app.register_required_components::<TnuaSubservientSensor, Position>();
         app.register_required_components::<TnuaSubservientSensor, Rotation>();
         app.register_required_components_with::<TnuaGravity, GravityScale>(|| GravityScale(0.0));
+    }
+}
+
+#[allow(clippy::type_complexity)]
+fn ensure_subservient_sensors_are_colliders_of_their_bevy_parents(
+    query: Query<(Entity, &ChildOf), (With<TnuaSubservientSensor>, Without<ColliderOf>)>,
+    mut commands: Commands,
+) {
+    for (entity, child_of) in query.iter() {
+        commands
+            .entity(entity)
+            // NOTE: Use the parent from child_of instead of `TnuaSubservientSensor::owner_entity`
+            // because Bevy parenting is how it works in Rapier and the sensor's owner is allowed
+            // to be different than the Bevy parent.
+            .insert_if_new(ColliderOf { body: child_of.0 });
     }
 }